new functionality added

+++ app/src/main/java/ru/gennady/kurbesov/simplesmartweather/model/WeatherState.kt
@@ -1,7 +1,7 @@
 package ru.gennady.kurbesov.simplesmartweather.model
 
 class WeatherState {
-    var id: Long? = null
+    var id: Int? = null
     var main: String? = null
     var description: String? = null
     var icon: String? = null
